feat: add `merge` configuration entry to `pull_request`
Allows optional merging of the auto-generated PR. Defaults to
'disabled'. User can configure merge strategy, branch deletion, the
merge commit message, and whether or not the PR should be merged
immediately, set to 'auto-merge' (merge as soon as requirements
are met), or if merge restrictions should be bypassed (assuming the
provided token has the required rights).

Overall implementation is based on the github CLI command
`gh pr merge`:

https://github.com/cli/cli/blob/e63dc2216b7eab552df9fc0dc2d3c5e045d461f7/pkg/cmd/pr/merge/merge.go

If the `commit` sub-key is left empty the GitHub provided default
message will be used.

Running the action with this option enabled will simply leave PRs that
are known to be unmanageable due to conflicts, repository policy, etc,
alone; however, if the merge of a PR that appears to be mergeable
fails, an error will be thrown.

Defaults:
---------
pull_request: {
  ...,
  merge: {
    mode: 'disabled',
    strategy: 'merge',
    delete_branch: false,
    commit: {},
  },
},

See README.md additions for more usage information.